# be/app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "RobotPet API",
        "description": "Flask API for RobotPet",
        "version": "1.0.0"
    },
    "host": "127.0.0.1:5000",
    "schemes": [
        "http"
    ],
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization",
            "description": "JWT 토큰을 포함한 Authorization 헤더"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/diary/summary": {
            "post": {
                "tags": [
                    "diary"
                ],
                "summary": "대화 내용 요약",
                "description": "사용자의 채팅방 ID를 기반으로 대화 내용을 요약하여 일기 형식으로 생성",
                "parameters": [
                    {
                        "name": "chatroom_id",
                        "in": "body",
                        "description": "대화 내용 요약을 위한 채팅방 ID",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "chatroom_id": {
                                    "type": "string",
                                    "example": "chatroom_123"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "대화 내용 요약 성공",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "summary": {
                                    "type": "string",
                                    "example": "오늘 대화는 모두 긍정적이었어요."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 (chatroom_id가 제공되지 않음)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "chatroom_id가 필요합니다."
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 내부 오류",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "서버 내부 오류"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/diary/save": {
            "post": {
                "tags": [
                    "diary"
                ],
                "summary": "일기 내용 저장",
                "description": "대화 내용을 요약하여 일기 형식으로 저장",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "user_id": {
                                    "type": "string",
                                    "example": "user_123"
                                },
                                "content": {
                                    "type": "string",
                                    "example": "오늘은 친구와 좋은 대화를 나누었어요."
                                },
                                "emotion": {
                                    "type": "string",
                                    "example": "happy"
                                },
                                "date": {
                                    "type": "string",
                                    "example": "2025-02-25"
                                }
                            },
                            "required": [
                                "user_id",
                                "content",
                                "emotion"
                            ]
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "일기 저장 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "일기 저장 성공!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 (필수 값 누락)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "필수 값이 누락되었습니다."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 내부 오류",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "서버 내부 오류"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/diary/summary/save": {
            "post": {
                "tags": [
                    "diary"
                ],
                "summary": "대화 내용 요약 및 저장",
                "description": "대화 내용 요약을 생성하고, 생성된 요약을 일기 형식으로 저장",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "chatroom_id": {
                                    "type": "string",
                                    "example": "chatroom_123"
                                }
                            },
                            "required": [
                                "chatroom_id"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "대화 내용 요약 및 저장 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "요약 및 일기 저장 성공!"
                                        },
                                        "summary": {
                                            "type": "string",
                                            "example": "오늘 대화는 모두 긍정적이었어요."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 (chatroom_id가 제공되지 않음)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "chatroom_id가 필요합니다."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 내부 오류",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "서버 내부 오류"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/diary/list": {
            "get": {
                "tags": [
                    "diary"
                ],
                "summary": "특정 날짜에 작성된 일기 목록 조회",
                "description": "사용자가 특정 날짜에 작성한 일기 목록을 조회하는 API",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "사용자의 고유 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "user_123"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "조회할 날짜 (형식: YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2025-02-23"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "일기 목록 조회 성공",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "user_123"
                                    },
                                    "date": {
                                        "type": "string",
                                        "example": "2025-02-23"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "오늘 하루도 잘 보냈어요. 기분이 좋았습니다."
                                    },
                                    "emotion": {
                                        "type": "string",
                                        "example": "happy"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "하루 종일 기분이 좋았고, 긍정적인 일들이 많았다."
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "잘못된 요청 (필수 값이 없거나 유효하지 않음)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "사용자 ID와 날짜는 필수입니다!"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 내부 오류",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "서버 내부 오류"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/diary/{diary_id}": {
            "get": {
                "tags": [
                    "diary"
                ],
                "summary": "일기 상세 조회",
                "description": "특정 일기의 상세 정보를 조회하는 API",
                "parameters": [
                    {
                        "name": "diary_id",
                        "in": "path",
                        "description": "상세 조회할 일기의 고유 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "diary_123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "일기 상세 조회 성공",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "user_id": {
                                    "type": "string",
                                    "example": "user_123"
                                },
                                "date": {
                                    "type": "string",
                                    "example": "2025-02-23"
                                },
                                "content": {
                                    "type": "string",
                                    "example": "오늘 하루도 잘 보냈어요. 기분이 좋았습니다."
                                },
                                "emotion": {
                                    "type": "string",
                                    "example": "happy"
                                },
                                "summary": {
                                    "type": "string",
                                    "example": "하루 종일 기분이 좋았고, 긍정적인 일들이 많았다."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "일기 없음 (해당 ID에 대한 일기 없음)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "일기 없음"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 내부 오류",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "서버 내부 오류"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/diary/delete/{diary_id}": {
            "delete": {
                "tags": [
                    "diary"
                ],
                "summary": "일기 삭제",
                "description": "특정 일기를 삭제하는 API",
                "parameters": [
                    {
                        "name": "diary_id",
                        "in": "path",
                        "description": "삭제할 일기의 고유 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "diary_123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "일기 삭제 성공",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "일기 삭제 성공!"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "일기 삭제 실패",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "일기 삭제 실패!"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 내부 오류",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "서버 내부 오류"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/diary/update/{diary_id}": {
            "put": {
                "tags": [
                    "diary"
                ],
                "summary": "일기 내용 수정",
                "description": "특정 일기의 내용을 수정하는 API",
                "parameters": [
                    {
                        "name": "diary_id",
                        "in": "path",
                        "description": "수정할 일기의 고유 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "diary_123"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "description": "새로운 일기 내용",
                                        "example": "오늘은 날씨가 정말 좋았다."
                                    },
                                    "emotion": {
                                        "type": "string",
                                        "description": "새로운 감정 상태",
                                        "example": "행복"
                                    }
                                },
                                "required": [
                                    "content"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "일기 수정 성공",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "일기 수정 성공!"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "일기 수정 실패",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "일기 내용은 필수입니다!"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 내부 오류",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "서버 내부 오류"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/diary/search": {
            "get": {
                "tags": [
                    "diary"
                ],
                "summary": "일기 내용 검색",
                "description": "특정 사용자 ID와 검색어를 기준으로 일기 내용을 검색하는 API",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "검색할 사용자의 고유 ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "user_123"
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "검색할 키워드",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "오늘"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "검색 결과 반환",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "diary_id": {
                                        "type": "string",
                                        "example": "diary_123"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "오늘은 날씨가 너무 좋았다. 기분도 좋았다."
                                    },
                                    "date": {
                                        "type": "string",
                                        "example": "2025-02-23"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "필수 입력값 누락",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "사용자 ID와 검색어는 필수입니다!"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버 내부 오류",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "서버 내부 오류"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}